Os write_alls e os read_alls estão a dar-me dor de cabeça

+++ proj_24-25-p1_base/common/constants.h
@@ -4,7 +4,7 @@
 #define MAX_PIPE_PATH_LENGTH 40 // tamanho max do caminho do pipe
 #define MAX_STRING_SIZE 40
 #define MAX_NUMBER_SUB 10
-#define MAX_CONNECT_MESSAGE_SIZE 121 //tamanho max da mensagem de conexao 
+#define MAX_CONNECT_MESSAGE_SIZE ((size_t)121)  //tamanho max da mensagem de conexao 
 #define MAX_CONNECT_RESPONSE_SIZE 3 //tamanho max da responsta da mensagem de conexao (op_code + char result + '\0')
 #define MAX_SUBSCRIBE_MESSAGE_SIZE 42 //tamanho max da mensagem de subscrição
 #define MAX_SUBSCRIBE_RESPONSE_SIZE 3 //tamanho max da resposta da mensagem de subscrição (op_code + char result + '\0)
